# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AlongRouteResolveRequest": {
        "additionalProperties": false,
        "properties": {
          "intent": {
            "const": "ADD_STOP_ALONG_ROUTE",
            "title": "Intent",
            "type": "string"
          },
          "provider_ref": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Provider Ref",
            "type": "string"
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "route": {
            "$ref": "#/components/schemas/RouteContextRequest"
          },
          "route_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Route Fingerprint",
            "type": "string"
          },
          "route_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Route Id",
            "type": "string"
          },
          "route_revision": {
            "minimum": 0.0,
            "title": "Route Revision",
            "type": "integer"
          },
          "session_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "intent",
          "session_id",
          "revision",
          "route_id",
          "route_revision",
          "route_fingerprint",
          "provider_ref",
          "route"
        ],
        "title": "AlongRouteResolveRequest",
        "type": "object"
      },
      "AlongRouteResolveResponse": {
        "additionalProperties": false,
        "properties": {
          "navigation_target": {
            "$ref": "#/components/schemas/compass__api__along_route_search__NavigationTargetResponse"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "route_id": {
            "title": "Route Id",
            "type": "string"
          },
          "route_revision": {
            "title": "Route Revision",
            "type": "integer"
          },
          "selection": {
            "$ref": "#/components/schemas/compass__api__along_route_search__ResolvedSelectionResponse"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "revision",
          "route_id",
          "route_revision",
          "selection",
          "navigation_target"
        ],
        "title": "AlongRouteResolveResponse",
        "type": "object"
      },
      "AlongRouteSearchRequest": {
        "additionalProperties": false,
        "properties": {
          "full_search": {
            "default": false,
            "title": "Full Search",
            "type": "boolean"
          },
          "intent": {
            "const": "ADD_STOP_ALONG_ROUTE",
            "title": "Intent",
            "type": "string"
          },
          "language": {
            "default": "it",
            "pattern": "^[a-z]{2}$",
            "title": "Language",
            "type": "string"
          },
          "page_cursor": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Cursor"
          },
          "query": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "route": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RouteContextRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "intent",
          "query",
          "session_id",
          "revision"
        ],
        "title": "AlongRouteSearchRequest",
        "type": "object"
      },
      "AlongRouteSearchResponse": {
        "additionalProperties": false,
        "properties": {
          "limitation": {
            "title": "Limitation",
            "type": "string"
          },
          "mode": {
            "enum": [
              "route_biased",
              "route_time_filtered",
              "global_specific"
            ],
            "title": "Mode",
            "type": "string"
          },
          "next_page_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Page Cursor"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SuggestionResponse"
            },
            "title": "Results",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "route_fingerprint": {
            "title": "Route Fingerprint",
            "type": "string"
          },
          "route_id": {
            "title": "Route Id",
            "type": "string"
          },
          "route_revision": {
            "title": "Route Revision",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "revision",
          "route_id",
          "route_revision",
          "route_fingerprint",
          "mode",
          "limitation",
          "next_page_cursor",
          "results"
        ],
        "title": "AlongRouteSearchResponse",
        "type": "object"
      },
...
      "BaseRouteRequest": {
        "additionalProperties": false,
        "properties": {
          "costing": {
            "const": "auto",
            "default": "auto",
            "title": "Costing",
            "type": "string"
          },
          "departure_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional timezone-aware departure instant. When omitted, traffic-aware routing departs now.",
            "title": "Departure At"
          },
          "destination": {
            "$ref": "#/components/schemas/CoordinateRequest"
          },
          "language": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z]{2}(-[A-Za-z]{2})?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "BCP 47 language tag used for maneuver instructions.",
            "title": "Language"
          },
          "origin": {
            "$ref": "#/components/schemas/CoordinateRequest"
          },
          "origin_heading_degrees": {
            "anyOf": [
              {
                "exclusiveMaximum": 360.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional observed travel heading at the origin. Used for a moving off-route recalculation so the new route departs in the current direction.",
            "title": "Origin Heading Degrees"
          },
          "origin_heading_tolerance_degrees": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Maximum angular difference accepted at the routed origin.",
            "title": "Origin Heading Tolerance Degrees"
          }
        },
        "required": [
          "origin",
          "destination"
        ],
        "title": "BaseRouteRequest",
        "type": "object"
      },
      "BaseRouteResponse": {
        "additionalProperties": false,
        "properties": {
          "distance_meters": {
            "minimum": 0.0,
            "title": "Distance Meters",
            "type": "number"
          },
          "duration_seconds": {
            "minimum": 0.0,
            "title": "Duration Seconds",
            "type": "number"
          },
          "geometry": {
            "$ref": "#/components/schemas/RouteGeometry"
          },
          "maneuvers": {
            "items": {
              "$ref": "#/components/schemas/ManeuverResponse"
            },
            "title": "Maneuvers",
            "type": "array"
          },
          "navigation": {
            "$ref": "#/components/schemas/NavigationTimingResponse"
          },
          "provider": {
            "const": "valhalla",
            "title": "Provider",
            "type": "string"
          },
          "speed_limit_source": {
            "anyOf": [
              {
                "const": "valhalla_graph",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Limit Source"
          },
          "speed_limits": {
            "items": {
              "$ref": "#/components/schemas/RouteSpeedLimitResponse"
            },
            "title": "Speed Limits",
            "type": "array"
          }
        },
        "required": [
          "distance_meters",
          "duration_seconds",
          "geometry",
          "maneuvers",
          "speed_limits",
          "speed_limit_source",
          "provider",
          "navigation"
        ],
        "title": "BaseRouteResponse",
        "type": "object"
      },
...
      "CoordinateRequest": {
        "additionalProperties": false,
        "properties": {
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "CoordinateRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaceSearchResponse": {
        "additionalProperties": false,
        "properties": {
          "cacheable": {
            "title": "Cacheable",
            "type": "boolean"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/PlaceSearchResultResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "cacheable",
          "results"
        ],
        "title": "PlaceSearchResponse",
        "type": "object"
      },
      "PlaceSearchResultResponse": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "display_name": {
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "kind": {
            "enum": [
              "address",
              "locality",
              "poi",
              "coordinate",
              "unknown"
            ],
            "title": "Kind",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/SearchCoordinateResponse"
          },
          "poi_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poi Name"
          },
          "provider": {
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          },
          "provider_place_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Place Id"
          },
          "result_id": {
            "minLength": 1,
            "title": "Result Id",
            "type": "string"
          }
        },
        "required": [
          "result_id",
          "display_name",
          "address",
          "location",
          "kind",
          "category",
          "poi_name",
          "provider",
          "provider_place_id"
        ],
        "title": "PlaceSearchResultResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OptionalHTTPBasic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CNG-aware navigation API foundations for Italy.",
    "title": "Compass CNG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/places/search": {
      "get": {
        "operationId": "place_search_api_v1_places_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 20,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "default": "it",
              "pattern": "^[A-Za-z]{2}(-[A-Za-z]{2})?$",
              "title": "Language",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid Compass API credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid search-provider response."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Search provider unavailable."
          }
        },
        "security": [
          {
            "OptionalHTTPBasic": []
          }
        ],
        "summary": "Place Search",
        "tags": [
          "place-search"
        ]
      }
    },
    "/api/v1/places/search-along-route": {
      "post": {
        "operationId": "search_along_route_api_v1_places_search_along_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlongRouteSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlongRouteSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid Compass API credentials."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OptionalHTTPBasic": []
          }
        ],
        "summary": "Search Along Route",
        "tags": [
          "place-search"
        ]
      }
    },
...
    "/api/v1/places/search-along-route/resolve": {
      "post": {
        "operationId": "resolve_along_route_api_v1_places_search_along_route_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlongRouteResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlongRouteResolveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid Compass API credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OptionalHTTPBasic": []
          }
        ],
        "summary": "Resolve Along Route",
        "tags": [
          "place-search"
        ]
      }
    },
    "/api/v1/routes": {
      "post": {
        "operationId": "base_route_api_v1_routes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaseRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseRouteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid Compass API credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No route was found."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid routing provider response."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Routing provider unavailable."
          }
        },
        "security": [
          {
            "OptionalHTTPBasic": []
          }
        ],
        "summary": "Base Route",
        "tags": [
          "routing"
        ]
      }
    },
...
  }
}